+++ docs/nuxt.config.js
@@ -2,17 +2,14 @@ import theme from '@nuxt/content-theme-docs'
 
 export default theme({
   i18n: {
-    locales: () => [{
-      code: 'es',
-      iso: 'es-ES',
-      file: 'es-ES.js',
-      name: 'Español'
-    }, {
-      code: 'en',
-      iso: 'en-US',
-      file: 'en-US.js',
-      name: 'English'
-    }],
+    locales: () => [
+      {
+        code: 'en',
+        iso: 'en-US',
+        file: 'en-US.js',
+        name: 'English'
+      }
+    ],
     defaultLocale: 'en'
   }
 })