+++ src/index.js
@@ -3,7 +3,7 @@ import { readdirSync } from 'fs'
 import merge from 'lodash.merge'
 // @ts-ignore
 import { directive as i18nExtensionsDirective } from '@intlify/vue-i18n-extensions'
-import { COMPONENT_OPTIONS_KEY, DEFAULT_OPTIONS, ROOT_DIR, STRATEGIES } from './helpers/constants'
+import { COMPONENT_OPTIONS_KEY, DEFAULT_OPTIONS, ROOT_DIR, STRATEGIES, REDIRECT_ON_OPTIONS } from './helpers/constants'
 import { buildHook, createExtendRoutesHook } from './core/hooks'
 import { formatMessage } from './templates/utils-common'
 
@@ -53,7 +53,8 @@ export default function (moduleOptions) {
   const templatesOptions = {
     Constants: {
       COMPONENT_OPTIONS_KEY,
-      STRATEGIES
+      STRATEGIES,
+      REDIRECT_ON_OPTIONS
     },
     nuxtOptions: {
       isUniversalMode: nuxtOptions.mode === 'universal',
@@ -68,10 +69,6 @@ export default function (moduleOptions) {
       continue
     }
     if (file.startsWith('plugin.')) {
-      if (file === 'plugin.seo.js' && !options.seo) {
-        continue
-      }
-
       this.addPlugin({
         src: resolve(templatesPath, file),
         fileName: join(ROOT_DIR, file),
@@ -93,6 +90,7 @@ export default function (moduleOptions) {
   this.nuxt.hook('build:before', () => buildHook.call(this, options))
 
   this.options.alias['~i18n-klona'] = require.resolve('klona/full').replace(/\.js$/, '.mjs')
+  this.options.alias['~i18n-ufo'] = require.resolve('ufo').replace(/\.js$/, '.mjs')
 
   if (!Array.isArray(this.options.router.middleware)) {
     throw new TypeError(formatMessage('options.router.middleware is not an array.'))