+++ src/templates/plugin.routing.js
@@ -1,7 +1,9 @@
 import './middleware'
 import Vue from 'vue'
 import { Constants, nuxtOptions, options } from './options'
-import { getDomainFromLocale } from './utils-common'
+import { getDomainFromLocale } from './plugin.utils'
+// @ts-ignore
+import { withoutTrailingSlash, withTrailingSlash } from '~i18n-ufo'
 
 /**
  * @this {import('../../types/internal').PluginProxy}
@@ -9,7 +11,7 @@ import { getDomainFromLocale } from './utils-common'
  */
 function localePath (route, locale) {
   const localizedRoute = resolveRoute.call(this, route, locale)
-  return localizedRoute ? localizedRoute.route.fullPath : ''
+  return localizedRoute ? localizedRoute.route.redirectedFrom || localizedRoute.route.fullPath : ''
 }
 
 /**
@@ -86,7 +88,7 @@ function resolveRoute (route, locale) {
       if (isPrefixed) {
         localizedRoute.path = `/${locale}${localizedRoute.path}`
       }
-      localizedRoute.path = localizedRoute.path.replace(/\/+$/, '') + (nuxtOptions.trailingSlash ? '/' : '') || '/'
+      localizedRoute.path = nuxtOptions.trailingSlash ? withTrailingSlash(localizedRoute.path, true) : withoutTrailingSlash(localizedRoute.path, true)
     }
   } else {
     if (!localizedRoute.name && !localizedRoute.path) {