+++ test/fixture/basic/layouts/default.vue
@@ -1,0 +1,11 @@
+<template>
+  <Nuxt />
+</template>
+
+<script>
+export default {
+  head () {
+    return this.$nuxtI18nHead({ addDirAttribute: false, addSeoAttributes: true })
+  }
+}
+</script>