+++ test/fixture/basic/pages/about.vue
@@ -4,6 +4,7 @@
     <div id="current-page">page: {{ $t('about') }}</div>
     <nuxt-link id="link-home" exact :to="localePath('index')">{{ $t('home') }}</nuxt-link>
     <div id="store-path-fr">{{ $store.state.routePathFr }}</div>
+    <div id="locale-properties-code">code: {{ code }}</div>
   </div>
 </template>
 
@@ -14,8 +15,15 @@ export default {
   components: {
     LangSwitcher
   },
+  /** @return {import('../../../../types/vue').NuxtI18nMeta} */
   head () {
     return this.$nuxtI18nHead()
+  },
+  computed: {
+    /** @return {string} */
+    code () {
+      return this.$i18n.localeProperties.code
+    }
   },
   nuxtI18n: {
     paths: {